Update MOIRE.cpp
Just adjusted the mini port to centre of red dot on the .svg. I had changed the .svg position also so the dot sits centre of the fader above. If you were to change both the x axis and the dot in the .svg the port position would be 33.6 iirc.