Update BAR.cpp
Hey seb. So I managed to get meters on the gain module working eventually. Scratching my head for a day or so as to why it was not showing up on make then today realised that it has to be named in the widget also, Doh. 

I was going for the colour scheme similar to ProTools meters red/orange/yellow/green probably a different type of meter. I only have green and red set at the moment. As it is, using green for rms and red for peak. Was thinking of doing green for low volume yellow for under -12dB orange over, red for approaching clipping etc.

Just a couple of things that are confusing. If I decrease the gain_param to -∞ the meters do not decrease they stick to the solid green/red colour but when I increase they work as expected. Any idea what would cause that? If I chain more gain modules together the RMS value goes lower and peak off the scale expected behaviour some what. I think this is how VCV handles overloaded dB not sure, masking. Not too worried about it overshooting I seen the same behaviour on bar, but I was thinking of adding a dB scale to the side as text so the low volume is important. Also thought of trying to add a fade curve eventually for low volume. 
The coordinates are really confusing eventually got a multiply for spacing. On bar this is 3 on my module 0.5. Have another version Gain and have it set to 35.31. What are those numbers thought they where pixels but did not make much sense being px.